fix: apply refactoring code
