+++ tee-worker/core/parentchain/indirect-calls-executor/src/indirect_calls/litentry/remove_identity.rs
